Add `experimental` key (close #112)
PR: #113

The Micro 1.3.2 release updated Stream Collector to 2.7.1. As part of the previous 2.7.0 release, a new key, `experimental` was added to the Stream Collector config. Without this key, the parsing of the config file fails, meaning Micro fails to start.

This has now been added to the Micro default collector config file.