Upgrade collector to 3.2.0 and enrich to 4.2.0
New collector and enrich versions mean:

* Cats Effect 3
* Http4s as web server, so no more Akka HTTP
* Using IO monad instead of Id

To be consistent with other Snowplow applications, this commit also brings small improvements to configuration loading.
Just like collector configuration, iglu resolver and all enrichments support HOCON configuration format.

As collector 3.2.x is on scala 2.13 and enrich 4.X on scala 2.12, it was easier to cross compile collector to 2.12 than upgrade enrich to 2.13. That's why micro is still on scala 2.12.